Implement read-only mode in Fractions page
Introduced a read-only mode for the Fractions page's InputNumberScroll component. A switch has been added to toggle between the editable and read-only state. In read-only mode, changes to the input value are prevented. Additional elements like an increment button have also been included for when the component is not in read-only mode.