Improve functionality and interaction of accordian control
This commit updates the accordian control to make it more user friendly and improve functionality. The changes include making the scroll behavior instant, improving decision logic when to show the accordion, and adding a minimum height. It also refactors the logic in the 'setItem' function and adds a 'required' attribute to the AccordianSelect component.