Remove try-catch blocks in RequestFormData and RequestObject functions
The try-catch blocks previously used in the RequestFormData and RequestObject functions in the FunctionsServer file have been removed. As a result, any errors that occur during the execution of these functions will now bubble up to the caller, enabling better error tracking and giving more responsibility for error handling to the higher-level functions that utilize these ones.